refactor: move Maneuver class to new maneuver.py
