+++ gmatpyplus/command/__init__.py
@@ -2,6 +2,7 @@ from .branch_command import *
 from .command import *
 from .end_target import *
 from .gmat_command import GmatCommand
+from .maneuver import Maneuver
 from .propagate import Propagate
 from .solver_branch_command import *
 from .solver_sequence_command import *
