+++ gmatpyplus/command/command.py
@@ -167,37 +167,6 @@ class EndFiniteBurn(GmatCommand):
 
         self.Initialize()
 
-
-class Maneuver(GmatCommand):
-    def __init__(self, name: str, burn: gp.ImpulsiveBurn | gp.FiniteBurn, spacecraft: gp.Spacecraft,
-                 backprop: bool = False):
-        """
-        Create a Maneuver command.
-
-        :param name:
-        :param burn:
-        :param spacecraft:
-        :param backprop:
-        """
-
-        super().__init__('Maneuver', name)
-
-        self.burn = burn
-        self.SetStringParameter(self.gmat_obj.GetParameterID('Burn'), self.burn.name)
-
-        self.spacecraft = spacecraft
-        self.SetStringParameter(self.gmat_obj.GetParameterID('Spacecraft'), self.spacecraft.name)
-        self.burn.SetSpacecraftToManeuver(self.spacecraft)  # update burn's assigned spacecraft
-
-        self.backprop = backprop
-        self.SetBooleanParameter(self.gmat_obj.GetParameterID('BackProp'), self.backprop)
-
-        self.SetSolarSystem()
-        self.SetObjectMap(gp.Moderator().GetConfiguredObjectMap())
-        self.SetGlobalObjectMap(gp.Sandbox().GetGlobalObjectMap())
-
-        self.Initialize()
-
 # class PropagateMulti(Propagate):
 #     # TODO: consider making this a nested/inner class of Propagate, so would call Propagate.Multi()
 #     """
